+++ tempodb/encoding/vparquet4/schema.go
@@ -55,6 +55,7 @@ const (
 	DefinitionLevelResourceSpansILSSpanEvent      = 4
 	DefinitionLevelResourceSpansILSSpanLink       = 4
 	DefinitionLevelResourceSpansILSSpanEventAttrs = 5
+	DefinitionLevelResourceSpansILSSpanLinkAttrs  = 5
 
 	FieldResourceAttrKey       = "rs.list.element.Resource.Attrs.list.element.Key"
 	FieldResourceAttrVal       = "rs.list.element.Resource.Attrs.list.element.Value.list.element"